setup points reduction when a barrel is destroyed
minus 50 points! It works pretty well, there are several optimizations we'll need to do on the destruction to improve things, but it's working!

Also disconnected the start widget for now since it's annoying while development is in progress, we can hook that up right at the end, but check out BP_startScreen if you want to take a peek!